+++ src/game-templates/2048/scripts/game_config_2048.gd
@@ -4,5 +4,5 @@ extends Resource
 
 @export var grid_size: int = 4
 @export var win_value: int = 2048
-@export var four_probability: float = 0.9  ## 出 2 的概率(1-x 出 4)
+@export var two_probability: float = 0.9  ## 出 2 的概率(1-x 出 4)
 @export var start_tiles: int = 2